requirements.txt
docxtpl>=0.16.0,<1.0
msal>=1.24.0,<2.0
requests>=2.31.0,<3.0
python-dotenv>=1.0.0,<2.0

# Dev/test
pytest>=7.0
pytest-cov
Pillow>=10.0.0
